check variations of name and replace them
