+++ components/cards/BasicMessages/BasicMessages.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useRef } from 'react';
+import { useRef } from 'react';
 import { STATUS } from '../../../enums';
 import CustomColorBtn from '../../CustomColorBtn/CustomColorBtn';
 import { INotification } from '../../Notification/Notification';
@@ -27,69 +27,7 @@ const BasicMessages: React.FC<IBasicMessages> = ({ messages, setMessages, notifi
         })
     };
 
-    function  globalName() {
-        const variations = [
-            /\[name\]/gi,
-            /\(name\)/gi,
-            /\{name\}/gi,
-            /\[nombre\]/gi,
-            /\(nombre\)/gi,
-            /\{nombre\}/gi,
-            /\[user\]/gi,
-            /\(user\)/gi,
-            /\{user\}/gi,
-            /\[usuario\]/gi,
-            /\(usuario\)/gi,
-            /\{usuario\}/gi,
-            // Add more variations here if needed
-          ];
-    }
 
-    
-
-    // -------------------------------------- //
-
-
-
-        const editableRef = useRef<HTMLDivElement>(null);
-      
-        // Function to set the cursor position at a specific index
-        const setCursorPosition = (index: number) => {
-
-        if (editableRef.current && editableRef.current.childNodes[0] && editableRef.current.childNodes[0].length > 0) {
-            const selection = window.getSelection();
-            const range = document.createRange();
-            const node = editableRef.current.childNodes[0]; // Assuming only one child node
-      
-            range.setStart(node, index);
-            range.collapse(true);
-            selection.removeAllRanges();
-            selection.addRange(range);
-          }
-        };
-      
-
-        function handleInput(e, index) {
-
-            let newMessage = messages[0]
-            newMessage[index] = e.currentTarget.innerText;
-            setMessages([newMessage])
-
-            const selection = window.getSelection();
-            if (selection) {
-            const range = selection.getRangeAt(0);
-            const offset = range.startOffset;
-            // console.log('Cursor position:', offset);
-            setTimeout(() => {
-                setCursorPosition(offset);
-            }, 10);
-            }
-        }
-
-
-        useEffect(() => {
-            console.log(messages)
-        }, [messages]);
 
     return (
             <div className={styles.basicMessages_cont}>
@@ -101,19 +39,12 @@ const BasicMessages: React.FC<IBasicMessages> = ({ messages, setMessages, notifi
                     {messages[0].map((message, index)=>{
                         return (
                             <div className={styles.messages_cont}>
-                                {/* <textarea value={message} placeholder={`Alternativa Nro. ${index + 1}`} onChange={(e)=>{
+                                <textarea value={message} placeholder={`Alternativa Nro. ${index + 1}`} onChange={(e)=>{
                                     let newMessage = messages[0]
                                     newMessage[index] = e.target.value;
                                     setMessages([newMessage])
-                                }} /> */}
-                                <p
-                                    ref={editableRef}
-                                    contentEditable
-                                    style={{ color: message.length > 0 ? "white" : "#ddd" }}
-                                    onInput={ (e) => { handleInput(e, index) } }
-                                >
-                                    {message.length > 0 ? message : `Alternativa Nro. ${index + 1}`}
-                                </p>
+                                }} />
+                                
                                 <img src="/close.svg" onClick={ 
                                     ()=>{
                                         if(messages[0].length > 1){
