+++ components/cards/Card/SendFree.tsx
@@ -14,6 +14,7 @@ import CardTitle from "../CardTitle/CardTitle";
 import { ContactInfo } from "../CardsContFree";
 import HeaderRow from "../HeaderRow/HeaderRow";
 import styles from "./FreeCard.module.css";
+import { globalName } from "./sendUtils";
 
 export interface IFreeCard3 {
   activeCard: number;
@@ -169,10 +170,13 @@ const FreeCard3: React.FC<IFreeCard3> = ({
 
     };
         
+    // Check [name] variations
+    let newCurrent = globalName(currentMessage)
+
     const sentMessage = await apiSenderWhatsappController.sendMessage(
       userInfo.user_id,
       destinatario.nombre,
-      currentMessage,
+      newCurrent,
       destinatario.numero,
       userInfo.access_token
     );
@@ -181,6 +185,7 @@ const FreeCard3: React.FC<IFreeCard3> = ({
     
 
   }catch(error){
+    setSendingState(SENDING_STATE.PAUSED)
     alert("Ocurrio un error inesperado en la plataforma. Por favor intenta mas tarde.")
   }
