+++ components/cards/CardsContFree.tsx
@@ -211,7 +211,7 @@ const CardsCont: React.FC<ICardsCont> = ({ isPaid }) => {
             default:
                 break;
         }
-    },[finalList, activeCard, messages])
+    },[finalList, activeCard, messages, sendingState])
 
 
     useEffect(() => {
